CHALLENGE 3
"The swarmbot communicates among itself and the need for human understanding of the bot must be performed with audio or visual indicators. This allows for the vehicles to act like vehicles that humans are used to using for transportation. Your team must ideate, conceptualize, design, build, test, and demonstrate functionality to enhance the human interaction behavior of the bot, and swarmbot community."

the bot follows the red line until it sees an obstacle, then changes lanes onto the yellow line, then continues it's journey until it stops at the wall.